Added code for implementing order of execution of construction files in an experiment, JUnit test for the same and changed return type of run method of simulateExperimentDirectory
Added a method in simulateExperimentDirectory to implement order of execution of construction files in an experiment taking care of dependencies between construction files (for example: if output of one construction file is an input to another, etc.)
JUnit test (CFOrderExecutionTest) was added to test the above functionality.
Changed return type of run method of simulateExperimentDirectory, the run method now returns an experiment object. Therefore FileDropGUI was modified as well.